refactor: move credential domain into ferriskey-domain
