+++ libs/ferriskey-domain/Cargo.toml
@@ -13,6 +13,7 @@ utoipa = { version = "5.4.0", features = ["chrono", "uuid"] }
 uuid = { version = "1.20.0", features = ["serde", "v4", "v7"] }
 thiserror = "2.0.12"
 tracing = "0.1"
+webauthn-rs = { version = "0.5.2", features = ["danger-credential-internals", "danger-allow-state-serialisation", "conditional-ui"] }
 mockall = { version = "0.14.0", optional = true }
 
 [dev-dependencies]
