+++ libs/ferriskey-domain/src/credential/entities.rs
