+++ libs/ferriskey-domain/src/credential/mod.rs
@@ -1,0 +1,3 @@
+pub mod entities;
+pub mod ports;
+pub mod services;
