+++ libs/ferriskey-domain/src/lib.rs
@@ -6,6 +6,8 @@ pub mod auth;
 pub mod authentication;
 pub mod client;
 pub mod common;
+pub mod credential;
+pub mod crypto;
 pub mod maintenance;
 pub mod realm;
 pub mod role;
