+++ libs/ferriskey-security/src/crypto/entities.rs
@@ -1,18 +1,4 @@
-#[derive(Debug, Clone)]
-pub struct HashResult {
-    pub hash: String,
-    pub salt: String,
-    pub hash_iterations: u32,
-    pub algorithm: String,
-}
-
-impl HashResult {
-    pub fn new(hash: String, salt: String, hash_iterations: u32, algorithm: String) -> Self {
-        Self {
-            hash,
-            salt,
-            hash_iterations,
-            algorithm,
-        }
-    }
-}
+//! `HashResult` moved to the shared `ferriskey-domain` crate so the `credential` domain (also in
+//! `ferriskey-domain`) can reference it without a `ferriskey-domain -> ferriskey-security` cycle.
+//! Re-exported here so `ferriskey_security::crypto::entities::HashResult` call sites keep resolving.
+pub use ferriskey_domain::crypto::HashResult;
